refactor: 合并 SelectProjectionContext 到 QueryContext
将投影查询逻辑从 SelectProjectionContext 子类合并到 QueryContext 主类中，
消除了不必要的继承层次，简化了查询上下文的结构。
